move components out of pages
