+++ web/src/features/riskMap/components/DayControl.tsx
