+++ web/src/features/riskMap/components/FireMap.tsx
@@ -23,7 +23,7 @@ import { AppDispatch } from '@/app/store'
 import CircleStyle from 'ol/style/Circle'
 import { FeatureCollection, MultiPolygon, Geometry, Point as GeoJSONPoint } from 'geojson'
 import { buffer, pointsWithinPolygon } from '@turf/turf'
-import { firePerimeterStyler } from '@/features/riskMap/pages/components/fireMapStylers'
+import { firePerimeterStyler } from '@/features/riskMap/components/fireMapStylers'
 import { ErrorBoundary } from '@/components'
 import { Point } from 'ol/geom'
 // import { parse } from 'csv-parse'
