+++ web/src/features/riskMap/components/fireMapStylers.ts
