+++ web/src/features/riskMap/components/fireSpreader.ts
