+++ web/src/features/riskMap/components/hotspot_points.json
