+++ web/src/features/riskMap/pages/RiskMapPage.tsx
@@ -1,5 +1,5 @@
-import { FireMap } from '@/features/riskMap/pages/components/FireMap'
-import { ValuesImportButton } from '@/features/riskMap/pages/components/UploadButton'
+import { FireMap } from '@/features/riskMap/components/FireMap'
+import { ValuesImportButton } from '@/features/riskMap/components/UploadButton'
 import { Box, CircularProgress, Grid, Modal, Typography } from '@mui/material'
 import VectorLayer from 'ol/layer/Vector'
 import VectorSource from 'ol/source/Vector'
@@ -8,13 +8,13 @@ import { Fill, Style } from 'ol/style'
 import { Map } from 'ol'
 import axios from 'api/axios'
 import { useEffect, useState } from 'react'
-import firePerimeterData from './components/PROT_CURRENT_FIRE_POLYS_SP.json'
-import hotspots from './components/FirespotArea_canada_c6.1_48.json'
-import { GrowFireButton } from '@/features/riskMap/pages/components/GrowFireButton'
+import firePerimeterData from '../components/PROT_CURRENT_FIRE_POLYS_SP.json'
+import hotspots from '../components/FirespotArea_canada_c6.1_48.json'
+import { GrowFireButton } from '@/features/riskMap/components/GrowFireButton'
 import { theme } from '@/app/theme'
 import { AppDispatch } from 'app/store'
 
-import DayControl from '@/features/riskMap/pages/components/DayControl'
+import DayControl from '@/features/riskMap/components/DayControl'
 import { useDispatch, useSelector } from 'react-redux'
 import { addGrowthLayer } from '@/features/riskMap/slices/fireGrowthSlice'
 import { selectFireGrowthDay } from '@/app/rootReducer'
